Conducts automated energy interactions analysis
Works with the Home directory. Requires the md.mdp file put there; this file must be WITHOUT the rows of energy groups.

The script requests the names it needs, outputting the questions to the terminal where it is running.

As GROMACS cannot work with tpr files having more than 64 energy groups, the script segments amino acid residues into a bunch of tpr's (which one contains fewer than 64 energy groups then).  

The basic principle of the script's work: the script generates an index file in a specific folder, copies the mdp-file, adds the required energy groups, leaves a bunch of tpr-s in the folder, deleting the mdp with the energy groups at the end. 

With every tpr then, it runs the analysis of energy terms printing results to the out_res.txt file situate in the Home directory.